create a new file for searchbar fields
